fix deplloy script

+++ deploy.sh
@@ -2,6 +2,6 @@
 
 set -e 
 
-rm -rf dist
+[ -d dist ] && rm -rf dist
 zola build --output-dir dist
 scp -r dist/* deployer@funky.studio:~/apps/static_websites/pavonz.com/
